refactor: move pubkeyToAddress to a src file
prep for loading from a bundle into XS
