+++ contract/test/test-xsnap-crypto.js
@@ -7,10 +7,7 @@ import proc from 'node:child_process';
 import fsMod from 'node:fs';
 import osMod from 'node:os';
 import tmpMod from 'tmp';
-import { fromBase64, toBech32 } from '@cosmjs/encoding';
-import { sha256 } from '@noble/hashes/sha256';
-
-import { ripemd160 } from '@noble/hashes/ripemd160';
+import { pubkeyToAddress } from '../src/check-sig.js';
 
 const { freeze } = Object;
 
@@ -61,19 +58,6 @@ export function options(io = {}) {
     messages,
   });
 }
-
-/**
- * @param {string} pubkey in base64
- * @param {string} prefix
- */
-const pubkeyToAddress = (pubkey, prefix) => {
-  const pubkeyData = fromBase64(pubkey);
-  assert.equal(pubkeyData.byteLength, 33);
-  //   console.log('pubkey', Buffer.from(pubkeyData));
-  const h1 = sha256.create().update(pubkeyData).digest();
-  const h2 = ripemd160.create().update(h1).digest();
-  return toBech32(prefix, h2);
-};
 
 test('pubkeyToAddress matches cosmjs tests', t => {
   const cases = [
